requirements.txt
Flask==2.3.3
flask-cors==3.0.10
openai==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0
httpx==0.27.2